Fix Wsign-compare warnings
For looping over a vector, use size_t.

There are a few instances where the index is appended to a vector of
unsigned ints, this should still work.

+++ source/AdmRenderer.cpp
@@ -172,7 +172,7 @@ namespace admrender {
 
 		// If the output layout has an LFE then get its index
 
-		for (int iSpk = 0; iSpk < m_outputLayout.channels.size(); ++iSpk)
+		for (size_t iSpk = 0; iSpk < m_outputLayout.channels.size(); ++iSpk)
 			if (!m_outputLayout.channels[iSpk].isLFE)
 				m_mapNoLfeToLfe.push_back(iSpk);
